Remove InnerManifestService by using interface default implementations
