+++ src/Microsoft.DotNet.ImageBuilder/src/Commands/PublishManifestCommand.cs
@@ -232,4 +232,3 @@ namespace Microsoft.DotNet.ImageBuilder.Commands
         }
     }
 }
-#nullable disable
