--- src/Microsoft.DotNet.ImageBuilder/src/IInnerManifestService.cs
@@ -1,14 +1,0 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-using System.Threading.Tasks;
-
-#nullable enable
-namespace Microsoft.DotNet.ImageBuilder
-{
-    public interface IInnerManifestService
-    {
-        Task<ManifestQueryResult> GetManifestAsync(string image, bool isDryRun);
-    }
-}
